Refactor: drop MoE EP=1 path; inline dispatch/combine into moe
- Remove dispatch.py and combine.py; move the distributed dispatch and
  combine definitions into moe.py, renaming N_LOCAL_EXPERTS/EP_WORLD_SIZE
  to moe's N_LOCAL/N_RANKS and adding the X_STAGE_ROWS constant.
- Delete the single-card EP=1 path (moe_ep1, moe_ep1_test,
  golden_moe_ep1, build_tensor_specs_ep1) and the *_ep1 dispatch/combine
  variants; --ep choices are now (2, 4, 8) and __main__ only runs the
  distributed l3_moe path.
- Remove the now-dead EP_ROUTING_GLOBAL flag: gate.py always routes over
  the full global expert set (N_EXPERTS = M.n_routed_experts), which moe
  shrinks to 32*EP before import, so the distributed path is unchanged.
- Drop the dead EP_RANK / EXPERTS_START_IDX constants left over from the
  removed EP=1 dispatch.
